feat: complete overview tab component
